namei: Fix cn_flags width in various places
This truncation is mostly harmless today, but fix it anyway to avoid
pain later down the road.

Approved by:	so
Reviewed by:	olce, kib
MFC after:	2 weeks
Differential Revision:	https://reviews.freebsd.org/D50417

(cherry picked from commit 0d224af399a66f00a5b33e5512fc018062cabf1d)
(cherry picked from commit 503ab844ceced16c7e8736c0caa586ff0bab5f0e)

+++ sys/kern/vfs_cache.c
@@ -4006,7 +4006,7 @@ SYSCTL_PROC(_vfs_cache_param, OID_AUTO, fast_lookup, CTLTYPE_INT|CTLFLAG_RW|CTLF
  */
 struct nameidata_outer {
 	size_t ni_pathlen;
-	int cn_flags;
+	uint64_t cn_flags;
 };
 
 struct nameidata_saved {
